+++ src/main/java/redis/clients/jedis/BinaryClient.java
@@ -36,7 +36,7 @@ public class BinaryClient extends Connection {
 
   private String password;
 
-  private int db;
+  private long db;
 
   private boolean isInWatch;
 
@@ -69,7 +69,7 @@ public class BinaryClient extends Connection {
     this.password = password;
   }
   
-  public void setDb(int db) {
+  public void setDb(long db) {
     this.db = db;
   }
 
@@ -918,7 +918,7 @@ public class BinaryClient extends Connection {
     sendCommand(GETRANGE, key, toByteArray(startOffset), toByteArray(endOffset));
   }
 
-  public int getDB() {
+  public Long getDB() {
     return db;
   }
